task 4

+++ python_async_function/3-tasks.py
@@ -2,8 +2,10 @@
 """asynchronous coroutine"""
 import asyncio
 
+
 wait_random = __import__("0-basic_async_syntax").wait_random
 
-def task_wait_random(max_delay: float) -> asyncio.Task:
+
+def task_wait_random(max_delay: int) -> asyncio.Task:
     """takes max_delay and returns a asyncio.Task"""
     return asyncio.create_task(wait_random(max_delay))
